+++ src/commands/general/synaptic.ts
@@ -1,9 +1,9 @@
-import { CommandDefinition } from '../../lib/command';
+import { CommandDefinition, replyWithMsg } from '../../lib/command';
 import { CommandCategory } from '../../constants';
 
 export const synaptic: CommandDefinition = {
     name: ['synaptic', 'syn'],
     description: 'Provides link to synaptic discord server',
     category: CommandCategory.GENERAL,
-    executor: (msg) => msg.channel.send('https://discord.gg/acQkSvrePG'),
+    executor: (msg) => replyWithMsg(msg, 'https://discord.gg/acQkSvrePG'),
 };