+++ src/commands/memes/ptu.ts
@@ -1,8 +1,8 @@
 import { CommandDefinition } from '../../lib/command';
-import { CommandCategory } from '../../constants';
+import { CommandCategory, imageBaseUrl } from '../../constants';
 import { makeEmbed } from '../../lib/embed';
 
-const PTU_URL = `${process.env.IMAGE_BASE_URL}memes/ptu.jpg`;
+const PTU_URL = `${imageBaseUrl}/memes/ptu.jpg`;
 
 export const ptu: CommandDefinition = {
     name: 'ptu',