+++ src/commands/support/utf-8.ts
@@ -1,13 +1,13 @@
 import { MessageCommandDefinition } from '../../lib/command';
-import { CommandCategory } from '../../constants';
+import { CommandCategory, imageBaseUrl } from '../../constants';
 import { makeEmbed, makeLines } from '../../lib/embed';
 
-const UTF8_HELP_URL = `${process.env.IMAGE_BASE_URL}support/utf-8.PNG`;
+const UTF8_HELP_URL = `${imageBaseUrl}/support/utf-8.PNG`;
 
 const utf8Embed = makeEmbed({
     title: 'FlyByWire A32NX | UTF-8',
     description: makeLines([
-        'Some users experience problems with various system in the A32NX. These are caused by an issue within MSFS which requires the use of the UTF8 Region setting in Windows.',
+        'Some users experience problems with various system in the A32NX. These are caused by an issue within MSFS which requires the use of the UTF-8 Region setting in Windows.',
         '',
         'To enable UTF-8 support follow the steps below:',
         '',